Core: Fixed compatibility of binary serialization GCC vs MSC (dword definition issue)
git-svn-id: svn://ultimatepp.org/upp/trunk@3637 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Painter/BufferPainter.h
@@ -126,7 +126,7 @@ private:
 	int                     min_y;
 	int                     max_y;
 	Size                    sz;
-	double                  mx;
+	int                     mx;
 
 	void  Init();
 	Cell *AddCells(int y, int n);
